backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pydantic
PyPDF2
python-docx
sentence-transformers
faiss-cpu
groq
